requirements.txt
mysql-connector-python
pandas
matplotlib
streamlit
